@@ -1,20 +1,20 @@
 name: cpx_research_sdk_flutter
 description: CPX Research SDK Flutter
-version: 1.0.2
+version: 1.1.0
 homepage: https://www.makeopinion.com
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^1.3.0
-  webview_flutter: ^4.10.0
+  http: ^1.5.0
+  webview_flutter: ^4.13.0
   json_annotation: ^4.9.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.4.13
-  json_serializable: ^6.9.0
+  build_runner: ^2.8.0
+  json_serializable: ^6.11.1
