@@ -24,7 +24,7 @@ final String IMAGE_URL = "dyn-image.cpx-research.com";
 class CPXNetworkService {
   // TODO: Update package version here as well
   // The package_info_plus package just shows the app version not the package version and adding the pubspec.yaml to the assets is a security issue for flutter web.
-  final String version = "1.0.2";
+  final String version = "1.1.0";
 
   CPXController controller = CPXController.controller;
   CPXData cpxData = CPXData.cpxData;
@@ -48,9 +48,9 @@ class CPXNetworkService {
     required CPXStyle style,
   }) {
     String backgroundColor =
-        ".${style.backgroundColor.value.toRadixString(16).substring(2)}";
+        ".${style.backgroundColor.toARGB32().toRadixString(16).substring(2)}";
     String textColor =
-        ".${style.textColor.value.toRadixString(16).substring(2)}";
+        ".${style.textColor.toARGB32().toRadixString(16).substring(2)}";
     Map<String, dynamic> params = _defaultRequestParameter;
     params['type'] = type.key;
     params['width'] = style.width.toString();
